+++ ImportBundle.php
@@ -12,6 +12,9 @@ namespace KimaiPlugin\ImportBundle;
 
 use App\Plugin\PluginInterface;
 use App\Validator\Constraints\TimesheetDeactivated;
+use App\Validator\Constraints\TimesheetLockdown;
+use App\Validator\Constraints\TimesheetLongRunning;
+use App\Validator\Constraints\TimesheetZeroDuration;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 class ImportBundle extends Bundle implements PluginInterface
@@ -23,5 +26,8 @@ class ImportBundle extends Bundle implements PluginInterface
         'kimai-timesheet-87', 'kimai-timesheet-88', 'kimai-timesheet-89', // timesheet deactivated before 2.0.16
         'kimai-timesheet-deactivated-activity', 'kimai-timesheet-deactivated-project', 'kimai-timesheet-deactivated-customer', // timesheet deactivated after 2.0.16
         TimesheetDeactivated::DISABLED_ACTIVITY_ERROR, TimesheetDeactivated::DISABLED_PROJECT_ERROR, TimesheetDeactivated::DISABLED_CUSTOMER_ERROR, // timesheet deactivated current codes
+        TimesheetLockdown::PERIOD_LOCKED, // lockdown is obviously not necessary to check here
+        TimesheetLongRunning::LONG_RUNNING, TimesheetLongRunning::MAXIMUM, TimesheetZeroDuration::ZERO_DURATION_ERROR, // if this was allowed in the past, it should be imported
+
     ];
 }