+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Compatibility: requires minimum Kimai 2.17.0
 
 - Make it simpler to increase max row and file size
 - Allow 4MB upload file size and 5000 rows per file
+- Deactivate timesheet validators during Kimai1 import: TimesheetLockdown, TimesheetLongRunning, TimesheetZeroDuration
 
 ## Version 2.13.0